services/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.4"
description = "PLOSolver backend Flask application"
authors = ["PLOSolver Team <hello@plosolver.com>"]
license = "All Rights Reserved"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.11.9"


# Core Flask dependencies
flask = "^3.1.1"
werkzeug = "^3.1.3"
flask-cors = "^6.0.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.6.0,<4.7.0"
flask-socketio = ">=5.5.1,<6.0.0"
python-socketio = "^5.10.0"

# gRPC dependencies
grpcio = "^1.62.0"
grpcio-tools = "^1.62.0"
grpcio-reflection = "^1.62.0"

# Database
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.0"

# Environment and configuration
python-dotenv = "^1.1.0"

# HTTP requests
requests = "^2.32.4"

# WSGI server
gunicorn = "^23.0.0"
eventlet = "^0.35.2"

# Payment processing
stripe = "^12.3.0"

# Authentication
authlib = "^1.6.0"
google-auth = "^2.40.3"
google-auth-oauthlib = "^1.2.2"
google-auth-httplib2 = "^0.2.0"

# Monitoring
prometheus-client = "^0.20.0"
psutil = "^6.0.0"

# Message queue
pika = "^1.3.2"
aio-pika = "^9.5.7"
redis = "^4.6.0"

# Vault integration
hvac = "^2.0.0"

# Celery for background tasks
celery = "^5.3.4"

# Core poker evaluation
treys = "^0.1.8"

# Machine learning dependencies (needed for solver)
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
joblib = "^1.3.0"
core = "1.1.3"

[tool.poetry.group.dev.dependencies]
# Testing dependencies
pytest = "^8.0.0"
pytest-flask = "^1.3.0"
pytest-mock = "^3.14.1"
pytest-cov = "^5.0.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.2"
responses = "^0.25.7"

# Core dependencies used by local plosolver_core when not installing from PyPI
deuces = "^0.2.0"

# Linting and code quality
flake8 = "^6.0.0"
black = "^23.0.0"
pre-commit = "^3.6.0"
isort = "^5.13.2"

# PostgreSQL testing dependencies
pytest-postgresql = "^5.0.0"
pytest-docker = "^3.0.0"

# Docker SDK for Python
docker = "^6.0.0"

# Data generation for tests
faker = "^18.0.0"
flake8-pyproject = "^1.2.3"
bandit = "^1.7.9"

[tool.poetry.group.build.dependencies]
# Build dependencies for Cython compilation
cython = "^3.0.0"
setuptools = "^61.0"
wheel = "^0.40.0"

[tool.poetry.urls]
"Homepage" = "https://github.com/plosolver/ploscope"
"Repository" = "https://github.com/plosolver/ploscope"

[[tool.poetry.source]]
name = "nexus-internal"
url = "https://nexus.ploscope.com/repository/pypi-internal/simple"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | src/protos
  | src/backend/protos
)/
'''

[tool.flake8]
max-line-length = 120
exclude = [
    "venv/",
    "__pycache__/",
    "*.pyc",
    "build_cython.py",
    "src/protos/",
    "src/backend/protos/"
]
ignore = ["E203", "W503", "E402"]



[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["src"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "docker: Docker-based tests"
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
